+++ crates/wdl-lint/CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated `RequirementsSection` to only check for missing `requirements` sections ([#1040](https://github.com/stjude-rust-labs/sprocket/pull/1040)).
 
+#### Removed
+
+* Removed the `ExceptDirectiveValid` rule ([#1125](https://github.com/stjude-rust-labs/sprocket/pull/1125)).
+
 ## 0.26.0 - 2026-08-05
 
 * New lint rule `InlineInstall` ([#1023](https://github.com/stjude-rust-labs/sprocket/pull/1023)).